Migrate webpack-dev-server config to v4 Add webpack base config and serve correct folder Require dependency from node_modules Add webpack base config and update dev-server options Update dependencies and fix conf Update cli version Migrate webpack-dev-server config to v4
