+++ components/automatic-grid/demo/demo.html
@@ -1,7 +1,7 @@
 <!--Copyright (c) Coherent Labs AD. All rights reserved. Licensed under the MIT License. See License.txt in the project root for license information. -->
 <!DOCTYPE html>
 <html lang="en">
-<link rel="stylesheet" href="../../grid/dist/grid.production.min.css">
+<link rel="stylesheet" href="../node_modules/coherent-gameface-grid/dist/grid.production.min.css">
 <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
 <link rel="stylesheet" href="../style.css">
