+++ tools/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "coherent-guic-cli",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "description": "A command line interface for creating Game UI Components for Gameface.",
   "main": "index.js",
   "scripts": {
@@ -20,8 +20,8 @@
     "rollup-plugin-multi-input": "1.1.1",
     "rollup-plugin-terser": "7.0.2",
     "webpack": "5.4.0",
-    "webpack-cli": "4.1.0",
-    "webpack-dev-server": "^4.3.1",
+    "webpack-cli": "4.7.0",
+    "webpack-dev-server": "^4.12.0",
     "yargs": "16.1.0"
   },
   "devDependencies": {
